major refactor: connection file split into 3 files to segregate bulk and transactional methods; added link to Anaplan Connec
